+++ package.json
@@ -55,6 +55,7 @@
     "react-dom": "^16.13.1",
     "react-hook-form": "^5.7.2",
     "react-image-file-resizer": "^0.3.8",
+    "react-infinite-scroll-component": "^6.1.0",
     "react-markdown": "^6.0.2",
     "react-mic": "^12.4.6",
     "react-redux": "^7.2.4",
