+++ src/components/molecules/Chatbox/Chatbox.scss
@@ -11,7 +11,16 @@
     flex-flow: column-reverse;
     justify-self: flex-end;
     padding: $spacing--xl $spacing--lg 0;
-    overflow-x: auto;
+    overflow-y: auto;
+  }
+
+  &__messages-infinite-scroll {
+    display: flex;
+    flex-flow: column-reverse;
+  }
+
+  &__messages-infinite-scroll-loading {
+    overflow: hidden;
   }
 
   &__form-box {
