+++ src/settings.ts
@@ -715,3 +715,5 @@ export const DEFAULT_TABLE_CAPACITY =
 export const ALLOWED_EMPTY_TABLES_NUMBER = 4;
 export const DEFAULT_JAZZBAR_TABLES_NUMBER = 12;
 export const DEFAULT_CONVERSATION_SPACE_TABLES_NUMBER = 10;
+
+export const CHATBOX_NEXT_RENDER_SIZE = 50;
